Relocate tests to new file
