+++ java/clients/src/test/java/sleeper/clients/util/table/TableWriterTest.java
@@ -18,11 +18,8 @@ package sleeper.clients.util.table;
 import org.junit.jupiter.api.Test;
 
 import sleeper.clients.testutil.ToStringConsoleOutput;
-import sleeper.core.properties.instance.CommonProperty;
-import sleeper.core.properties.instance.InstanceProperty;
 
 import java.io.IOException;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.clients.testutil.ClientTestUtils.example;
@@ -160,70 +157,5 @@ class TableWriterTest {
                 .build().write(output.getPrintStream());
         // Then
         assertThat(output).hasToString(example("reports/table/markdown.txt"));
-    }
-
-    @Test
-    void shouldCreateMarkdownTableWithProvidedInstanceProperties() throws IOException {
-        // Given
-        TableWriterFactory.Builder factoryBuilder = TableWriterFactory.builder().structure(TableStructure.MARKDOWN_FORMAT);
-        factoryBuilder.addFields(TableWriterPropertyHelper.getMarkdownFields());
-
-        TableWriterFactory factory = factoryBuilder.build();
-        ToStringConsoleOutput output = new ToStringConsoleOutput();
-
-        InstanceProperty property1 = CommonProperty.JARS_BUCKET;
-        InstanceProperty property2 = CommonProperty.FILE_SYSTEM;
-
-        // When
-        factory.tableBuilder()
-                .row(TableWriterPropertyHelper.generatePropertyDetails(factoryBuilder.getFields(), property1))
-                .row(TableWriterPropertyHelper.generatePropertyDetails(factoryBuilder.getFields(), property2))
-                .build().write(output.getPrintStream());
-        // Then
-        assertThat(output).hasToString(example("reports/table/property.txt"));
-    }
-
-    @Test
-    void shouldCreateMarkdownTableWithLongDescriptionEntry() throws IOException {
-        // Given
-        TableStructure structure = TableStructure.MARKDOWN_FORMAT;
-        TableWriterFactory.Builder factoryBuilder = TableWriterFactory.builder().structure(structure);
-        factoryBuilder.addFields(TableWriterPropertyHelper.getMarkdownFields());
-
-        TableWriterFactory factory = factoryBuilder.build();
-        ToStringConsoleOutput output = new ToStringConsoleOutput();
-
-        InstanceProperty property1 = CommonProperty.TASK_RUNNER_LAMBDA_MEMORY_IN_MB;
-        InstanceProperty property2 = CommonProperty.TASK_RUNNER_LAMBDA_TIMEOUT_IN_SECONDS;
-
-        // When
-        factory.tableBuilder()
-                .row(TableWriterPropertyHelper.generatePropertyDetails(factoryBuilder.getFields(), property1))
-                .row(TableWriterPropertyHelper.generatePropertyDetails(factoryBuilder.getFields(), property2))
-                .build().write(output.getPrintStream());
-        // Then
-        assertThat(output).hasToString(example("reports/table/propertyLong.txt"));
-
-    }
-
-    @Test
-    void shouldCreateValidPropertiesTableWhenGivenAStreamOfProperties() throws IOException {
-        // Given
-        TableStructure structure = TableStructure.MARKDOWN_FORMAT;
-        TableWriterFactory.Builder factoryBuilder = TableWriterFactory.builder().structure(structure);
-        factoryBuilder.addFields(TableWriterPropertyHelper.getMarkdownFields());
-
-        ToStringConsoleOutput output = new ToStringConsoleOutput();
-
-        List<InstanceProperty> propertyList = List.of(CommonProperty.TASK_RUNNER_LAMBDA_MEMORY_IN_MB,
-                CommonProperty.TASK_RUNNER_LAMBDA_TIMEOUT_IN_SECONDS);
-
-        TableWriter writer = TableWriterPropertyHelper.generateTableBuildForGroup(propertyList.stream());
-
-        // When
-        writer.write(output.getPrintStream());
-        // Then
-        assertThat(output).hasToString(example("reports/table/propertyLong.txt"));
-
     }
 }
